Upgrade deps; refactor MemInode => MemInodeBase<Derived>.
