+++ src/llfs/mem_inode.cpp
@@ -77,136 +77,6 @@ auto MemInode::lookup_child(const std::string& name) -> batt::StatusOr<const fus
 
 //==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
 //
-auto MemInode::forget(u64 count) -> IsDead
-{
-  return this->remove_lookup(count);
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void MemInode::add_lookup(usize count) noexcept
-{
-  const u64 increment = MemInode::kLookupCountIncrement * count;
-  const u64 prior_value = this->count_.fetch_add(increment);
-
-  BATT_CHECK_LT(MemInode::get_lookup_count(prior_value), MemInode::kMaxLookupCount);
-  BATT_CHECK_LT(MemInode::get_lookup_count(prior_value),
-                MemInode::get_lookup_count(prior_value + increment));
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-auto MemInode::remove_lookup(usize count) noexcept -> IsDead
-{
-  const u64 increment = MemInode::kLookupCountIncrement * count;
-  const u64 prior_value = this->count_.fetch_sub(increment);
-
-  BATT_CHECK_GE(MemInode::get_lookup_count(prior_value), count)
-      << BATT_INSPECT(prior_value) << BATT_INSPECT(count) << BATT_INSPECT(increment)
-      << BATT_INSPECT(std::bitset<64>{kLookupCountMask}) << BATT_INSPECT(kLookupCountShift);
-
-  BATT_CHECK_GT(MemInode::get_lookup_count(prior_value),
-                MemInode::get_lookup_count(prior_value - increment))
-      << BATT_INSPECT(prior_value) << BATT_INSPECT(count) << BATT_INSPECT(increment);
-
-  return MemInode::is_dead_state(prior_value - increment);
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-batt::Status MemInode::increment_link_refs(usize count) noexcept
-{
-  const u64 increment = MemInode::kLinkCountIncrement * count;
-  const u64 prior_value = this->count_.fetch_add(increment);
-
-  BATT_CHECK_LT(MemInode::get_link_count(prior_value), MemInode::kMaxLinkCount);
-  BATT_CHECK_LT(MemInode::get_link_count(prior_value),
-                MemInode::get_link_count(prior_value + increment));
-
-  if ((prior_value & MemInode::kDeadFlag) != 0) {
-    this->count_.fetch_sub(increment);
-    return batt::status_from_errno(ENOENT);
-  }
-
-  return batt::OkStatus();
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-auto MemInode::decrement_link_refs(usize count, RequireEmpty require_empty) noexcept
-    -> batt::StatusOr<IsDead>
-{
-  if (count == 0) {
-    return {batt::status_from_errno(EINVAL)};
-  }
-
-  const u64 increment = MemInode::kLinkCountIncrement * count;
-
-  BATT_REQUIRE_OK(this->acquire_count_lock());
-  auto on_scope_exit = batt::finally([&] {
-    this->release_count_lock();
-  });
-
-  const batt::Optional<u64> updated_value =
-      this->count_.modify_if([&](u64 observed_value) -> batt::Optional<u64> {
-        const u64 observed_link_count = MemInode::get_link_count(observed_value);
-        const bool will_be_dead = ((observed_link_count - count) == 0);
-
-        BATT_CHECK_LE(count, observed_link_count);
-        BATT_CHECK_GE(MemInode::get_link_count(observed_value),
-                      MemInode::get_link_count(observed_value - increment))
-            << "Integer wrap!";
-
-        // If the requested decrement would bring the link count to zero but we are not empty, this
-        // is an error.
-        //
-        if (require_empty && will_be_dead && !this->is_empty()) {
-          return batt::None;
-        }
-
-        // Everything looks good!  Attempt to CAS-modify the count.
-        //
-        u64 target_value = observed_value - increment;
-        if (will_be_dead) {
-          target_value |= MemInode::kDeadFlag;
-        }
-
-        return target_value;
-      });
-
-  if (!updated_value) {
-    return {batt::status_from_errno(ENOTEMPTY)};
-  }
-
-  return MemInode::is_dead_state(*updated_value);
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-batt::Status MemInode::acquire_count_lock() noexcept
-{
-  for (;;) {
-    const u64 prior_value = this->count_.fetch_or(MemInode::kLockFlag);
-    if (!(prior_value & MemInode::kLockFlag)) {
-      break;
-    }
-    if ((prior_value & MemInode::kDeadFlag)) {
-      return {batt::status_from_errno(ENOENT)};
-    }
-    batt::Task::yield();
-  }
-  return batt::OkStatus();
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
-void MemInode::release_count_lock() noexcept
-{
-  this->count_.fetch_and(~MemInode::kLockFlag);
-}
-
-//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
-//
 auto MemInode::get_attributes() -> batt::StatusOr<FuseImplBase::Attributes>
 {
   return {FuseImplBase::Attributes{
